[DDW-821] Moves shared-helpers and screenshot helpers into new Common domain
